refactor control->controller
